refactor: move tv classes to tv folder
