+++ src/parser/classes/Tab.ts
@@ -4,7 +4,7 @@ import SectionList from './SectionList.js';
 import MusicQueue from './MusicQueue.js';
 import RichGrid from './RichGrid.js';
 import { YTNode } from '../helpers.js';
-import TvSurfaceContent from './TvSurfaceContent.js';
+import TvSurfaceContent from './tv/TvSurfaceContent.js';
 
 export default class Tab extends YTNode {
   static type = 'Tab';
