+++ src/parser/classes/tv/TvSecondaryNav.ts
@@ -1,7 +1,7 @@
-import { YTNode, type ObservedArray } from '../helpers.js';
-import { Parser, type RawNode } from '../index.js';
+import { YTNode, type ObservedArray } from '../../helpers.js';
+import { Parser, type RawNode } from '../../index.js';
 import TvSecondaryNavSection from './TvSecondaryNavSection.js';
-import Text from './misc/Text.js';
+import Text from '../misc/Text.js';
 
 export default class TvSecondaryNav extends YTNode {
   static type = 'TvSecondaryNav';
