+++ src/parser/yttv/SubscriptionsFeed.ts
@@ -6,7 +6,7 @@ import type { ObservedArray, YTNode } from '../helpers.js';
 import { observe } from '../helpers.js';
 import { InnertubeError } from '../../utils/Utils.js';
 import Tab from '../classes/Tab.js';
-import TvSurfaceContent from '../classes/TvSurfaceContent.js';
+import TvSurfaceContent from '../classes/tv/TvSurfaceContent.js';
 import Grid from '../classes/Grid.js';
 
 export default class SubscriptionsFeed {
